% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coltrans.R
\name{coltrans}
\alias{coltrans}
\title{Transform column types}
\usage{
coltrans(file)
}
\arguments{
\item{file}{The file you want to transform the columns of}
}
\value{
Tibble with changed column types
}
\description{
Transform the column types of the tidy flu_data and dengue_data in my online portfolio page, chapter: relational databases
}
