% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_data.R
\name{mean_data}
\alias{mean_data}
\title{Calculate the mean of Rawdata in the portfolio}
\usage{
mean_data(data)
}
\arguments{
\item{data}{The dataset to calculate the mean of}
}
\value{
Updated table with mean values
}
\description{
calculate the mean value of the Rawdata column in my online portfolio page, chapter: C elegans Plate Experiment
}
