% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot data using ggplot}
\usage{
plot_data(data)
}
\arguments{
\item{data}{The data to be plotted}
}
\value{
Scatterplot graph
}
\description{
Plot data in a scatterplot using ggplot for 2 plots in my online portfolio page, chapter: C elegans Plate Experiment
}
