% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coltrans.R
\name{coltrans}
\alias{coltrans}
\title{Transform columntypes}
\usage{
coltrans(x)
}
\arguments{
\item{x}{= File you want to transform the columns of}
}
\value{
Tibble with changed column types
}
\description{
Transform the columntypes of the tidy flu_data and dengue_data in my online portfolio page: relational databases
}
