% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatrans.R
\name{datatrans}
\alias{datatrans}
\title{Transform data}
\usage{
datatrans(x)
}
\arguments{
\item{x}{= file you want to transform the data of}
}
\value{
Tibble with altered data
}
\description{
Transform data in flu_data and dengue_data to make them joinable
}
